Move rejection_gamma, add gamma_dirichlet
