+++ pyro/distributions/testing/fakes.py
@@ -1,12 +1,17 @@
 from __future__ import absolute_import, division, print_function
 
-from pyro.distributions.beta import Beta
-from pyro.distributions.gamma import Gamma
-from pyro.distributions.normal import Normal
+from pyro.distributions.torch.beta import Beta
+from pyro.distributions.torch.dirichlet import Dirichlet
+from pyro.distributions.torch.gamma import Gamma
+from pyro.distributions.torch.normal import Normal
 from pyro.distributions.random_primitive import RandomPrimitive
 
 
 class NonreparameterizedBeta(Beta):
+    reparameterized = False
+
+
+class NonreparameterizedDirichlet(Dirichlet):
     reparameterized = False
 
 
@@ -19,5 +24,6 @@ class NonreparameterizedNormal(Normal):
 
 
 nonreparameterized_beta = RandomPrimitive(NonreparameterizedBeta)
+nonreparameterized_dirichlet = RandomPrimitive(NonreparameterizedDirichlet)
 nonreparameterized_gamma = RandomPrimitive(NonreparameterizedGamma)
 nonreparameterized_normal = RandomPrimitive(NonreparameterizedNormal)
