+++ pyro/distributions/testing/rejection_gamma.py
