+++ setup.cfg
@@ -1,10 +1,12 @@
 [flake8]
 max-line-length = 120
-exclude = docs/src, examples/storyboard
+exclude = docs/src
 
 [isort]
 line_length = 120
 not_skip = __init__.py
+known_first_party = pyro, tests
+known_third_party = six, torch
 
 [tool:pytest]
 filterwarnings = error
